+++ CHANGES.md
@@ -45,11 +45,15 @@ TODO: this is not finished yet:
     ```
 * `GroundPrimitive` now supports batching geometry for better performance.
 * Added `BoundingSphere.fromEncodedCartesianVertices` to create bounding volumes from parallel arrays of the upper and lower bits of `EncodedCartesian3`s.
+* Added `ImageryLayer.getViewableRectangle` to make it easy to get the effective bounds of an imagery layer.
 * Fixed creating bounding volumes for `GroundPrimitive`s whose containing rectangle has a width greater than pi.
 * Fixed incorrect texture coordinates for polygons with large height.
 * Fixed camera.flyTo not working when in 2D mode and only orientation changes
 * Added `UrlTemplateImageryProvider.reinitialize` for changing imagery provider options without creating a new instance.
 * `UrlTemplateImageryProvider` now accepts a promise to an `options` object in addition to taking the object directly.
+* Added a Sandcastle example to "star burst" overlapping billboards and labels.
+* Added `Rectangle.union` and `Rectangle.expand`.
+* Added `Math.logBase` function.
 
 ### 1.17 - 2016-01-04